% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{add_times}
\alias{add_times}
\title{Add time series columns}
\usage{
add_times(d_in)
}
\description{
Add time series columns
}
\examples{
data(belle)
belle <- add_times(belle)
cols <- c("ID", "Time", "hr", "day", "dayhr")
## belle <- data.frame(belle)
head(belle[, cols])
## belle[, ..cols]
## tibble::as_tibble(belle)
}
