% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSSBperR.R
\name{calc_ssb_per_r}
\alias{calc_ssb_per_r}
\title{Calculate spawning stock biomass per recruit}
\usage{
calc_ssb_per_r(nAge, ssbwaa, maturity, maa, fmult, selx, spawntime)
}
\arguments{
\item{nAge}{integer number of ages}

\item{ssbwaa}{vector of spawning stock biomass weights at age}

\item{maturity}{vector of maturity at age}

\item{maa}{vector of natural mortality at age}

\item{fmult}{double fishing mortality multiplier}

\item{selx}{vectory of fishery selectivity at age}

\item{spawntime}{double fraction of year before spawning occurs (should be within [0, 1])}
}
\value{
spr double value of spawners per recruit
}
\description{
Calculate spawning stock biomass per recruit
}
