% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSSBrho.R
\name{calcSSBrho}
\alias{calcSSBrho}
\title{Calculates Mohn's rho for spawning stock biomass from ASAP retrospective}
\usage{
calcSSBrho(fname, n.peels)
}
\arguments{
\item{fname}{character path and name of ASAP input file (including .dat extension)}

\item{n.peels}{integer number of peels used in calculating mean Mohn's rho value (must be <10)}
}
\value{
double mean Mohn's rho for SSB
}
\description{
Computes relative change in SSB from tip of previous n.peels relative to terminal estimates
for that year. Returns the average value.
}
