% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_encounter}
\alias{check_encounter}
\title{Check the encounter points}
\usage{
check_encounter(data0, data1)
}
\arguments{
\item{data0}{dped}

\item{data1}{dveh}
}
\value{
a NULL or a string
}
\description{
Check the encounter points
}
\examples{
is_encounter <- check_encounter(dped, devh)

}
