% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_hotzone}
\alias{check_hotzone}
\title{Check whether the encounter point was in a marked crossing.}
\usage{
check_hotzone(data, xy, site = "belle", verbose = FALSE)
}
\arguments{
\item{data}{a data frame a dveh or dped}

\item{xy}{the encounter point found by find_encounter(dped, dveh)}

\item{site}{a string to differentiate Belle Isle Road from Queensway}

\item{verbose}{print a useful message}

\item{zone}{a scalar, representing 1, 2, 3 or 4 for the four crossing 
zones in Belle Isle Road. 5, 6, 7, or 8 for the four crossing zones in 
Queensway.}
}
\value{
a NULL or a data frame
}
\description{
The numeric x-y values for the zone polygon must match the png file. It
appears that Viscando provides slight different png files, even for the 
same investigative site. The street image with the coordinate system matches
the numerical X-Y value in the data Viscando provided.
}
\examples{
hotzone <- data.table(x=c(10, 10, 130,     130,   300, 300, 450, 450, 
150, 150, 270,   270,   310, 310, 420, 420,
780, 780, 1100, 1100,  1300, 1300, 1650, 1650,
1400, 1400, 1700, 1700), 
y=c(120, 260, 260, 120,   150, 300, 300, 150,
    350, 520, 520, 350,   350, 520, 520, 350,
    600, 1000, 1000, 600, 50, 300, 300, 50,
    380, 700, 700, 380),
value = rep(c(1, 2, 3, 4, 5, 6, 7), each = 4),
z = rep(c("um", "um", "m", "m", "m", "um", "um"),  each = 4),
site = c(rep("belle", 16), rep("qw", 12)))

}
