% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{checkSpeed}
\alias{checkSpeed}
\title{Check whether a car is slowing down}
\usage{
checkSpeed(data, idx, zone = NULL, rate = 2, verbose = FALSE)
}
\arguments{
\item{data}{dveh}

\item{rate}{the rate of slowing down}

\item{verbose}{printing messages}

\item{critical}{index}
}
\value{
a Boolean
}
\description{
Check whether a car is slowing down
}
\examples{
## 

}
