% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanUpFiles.R
\name{cleanUpFiles}
\alias{cleanUpFiles}
\title{Remove files created during runRetroMults}
\usage{
cleanUpFiles(fname)
}
\arguments{
\item{fname}{character base file name used in runRetroMults}
}
\description{
Deletes files in current directory associated with runRetroMults analyses. Removes both 
named files as well as default files created when ASAP runs.
}
