% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extract_data}
\alias{extract_data}
\title{Extract data from all csv files}
\usage{
extract_data(fpath, is_print = TRUE)
}
\arguments{
\item{fpath}{file path}

\item{is_print}{preview the first few lines of the data}
}
\description{
This function was to aggregate all csv files into one data.table
}
