% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{find_encounter}
\alias{find_encounter}
\title{Find the closest point between two agents}
\usage{
find_encounter(dat0, dat1)
}
\value{
a vector storing the coordinate a vehicle (dat1) and a pedestrian
(dat0)
}
\description{
The function scans over all time point for a agent 1 to look for the 
closest encounter point in a agent 2, ignoring whether they are at the 
same time points
}
