% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{find_video}
\alias{find_video}
\title{A tool for finding which video to watch}
\usage{
find_video(data, rng, which_site = 1, verbose = TRUE)
}
\arguments{
\item{data}{a standard Viscando data frame.}

\item{rng}{extending the end range.}

\item{which_site}{default to site 1}
}
\value{
a string
}
\description{
Must work with find_range.
}
\examples{
## TODO update the example
load("tests/extdata/day2/pd.rda")
d <- data.table(cbind(x1, pet, zoneid))
j <- d[1, ]$ped
k <- d[1, ]$veh
c(j, k)
fndat <- paste0("tests/tmp_data/day2/p",j, "-v",k, ".rda")
load(fndat) 
tmp <- find_video(dped, 3)

}
