% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getASAPrecentVals.R
\name{get_asap_recent_vals}
\alias{get_asap_recent_vals}
\title{Extract recent averages of ASAP estimates for per recruit and projection calculations.}
\usage{
get_asap_recent_vals(asap, nrecentyears)
}
\arguments{
\item{asap}{list from reading asap.rdat file using dget()}

\item{nrecentyears}{integer number of recent years to include in averages}
}
\value{
list of values
}
\description{
Extract recent averages of ASAP estimates for per recruit and projection calculations.
}
