% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStartingNAA.R
\name{get_starting_naa}
\alias{get_starting_naa}
\title{Compute numbers at age in terminal year plus one and fill in recruitment using median of input years.}
\usage{
get_starting_naa(asap, recruityears)
}
\arguments{
\item{asap}{list from reading asap.rdat file using dget()}

\item{recruityears}{vector of years to include when computing median recruitment}
}
\value{
vector of numbers at age
}
\description{
Compute numbers at age in terminal year plus one and fill in recruitment using median of input years.
}
