% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{m2pixel}
\alias{m2pixel}
\alias{p2metre}
\title{Check whether a car is slowing down}
\usage{
m2pixel(x, isX = TRUE, site = "Belle Road")

p2metre(input, isX = TRUE, site = "belle")
}
\arguments{
\item{isX}{either X or Y}

\item{site}{either which investigative sites Belle Road or QW.}

\item{X}{X in metre}
}
\description{
Check whether a car is slowing down
}
\examples{
## 

}
