% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_simple_traj}
\alias{plot_simple_traj}
\title{Plot a simple trajectory}
\usage{
plot_simple_traj(
  start,
  finish,
  start_next = NULL,
  finish_next = NULL,
  xlim = c(-50, 50),
  ylim = c(-50, 50),
  continue = FALSE
)
}
\description{
Plot a simple trajectory
}
