% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectShortTerm.R
\name{project_short_term}
\alias{project_short_term}
\title{Compute catch in short term deterministic projections.}
\usage{
project_short_term(naa, maa, fmult, selx, cwaa, nprojyears, recruits)
}
\arguments{
\item{naa}{vector of starting numbers at age}

\item{maa}{vector of natural mortality at age}

\item{fmult}{double fishing mortality multiplier}

\item{selx}{vector of fishery selectivity}

\item{cwaa}{vector of catch weight at age}

\item{nprojyears}{integer number of years in projection}

\item{recruits}{double recruitment in projection year 2 and onwards}
}
\value{
vector of catch in weight for nprojyears
}
\description{
Compute catch in short term deterministic projections.
}
