% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{records}
\alias{records}
\title{Which Days We Recorded the Data}
\usage{
records()
}
\value{
a matrix
}
\description{
See \file{doc-4947975-1.0-Belle Isle Road - Leeds.pdf} and 
\file{doc-4947976-1.0-Queensway - Leeds.pdf} for further details.
}
\details{
\enumerate{
    \item Belle: From May 9, 2022, 08:50 PM to May 15, 2022, 08:05 AM
    \item Qw: From May, 17, 2022, 11:17 AM to May, 23, 2022, 04:05 PM
 }
}
\examples{
per <- records()
}
