% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRetroMults.R
\name{runRetroMults}
\alias{runRetroMults}
\title{Wrapper to run ASAP retrospective across range of multipliers for catch, M, or indices}
\usage{
runRetroMults(scenario.name, asap.fname, n.peels, ramp, year.vals,
  cmult.vals, mmult.vals, mselx, mselxlab, imult.vals,
  save.files = FALSE)
}
\arguments{
\item{scenario.name}{character describes the runs in words to distinguish from other sets of runs}

\item{asap.fname}{character name of asap dat file (including extension)}

\item{n.peels}{integer number of peels used to compute Mohn's rho}

\item{ramp}{integer number of years in ramp (0 means immediate jump to multiplied value)}

\item{year.vals}{vector of change years}

\item{cmult.vals}{vector of catch multipliers}

\item{mmult.vals}{vector of natural mortality (M) multipliers}

\item{mselx}{vector of selectivity at age to mutiply mmult by (allows additional mortality to be for all ages equally, focused on young ages, focused on old ages, etc.)}

\item{mselxlab}{character describes natural mortality selectivity used}

\item{imult.vals}{vector of index multipliers}

\item{save.files}{logical TRUE means keep all retrospective files, default FALSE removes all files}
}
\value{
data frame of scenario, change year, ramp, catch multiplier, M multiplier, M selectivity label, and Mohn's rho for SSB
}
\description{
Scenarios defined by ramp (number of years included in ramp of multipliers), change year, 
catch multipliers, natural mortality multipliers, natural mortality selectivty 
(maximum value should be 1, minimum value allowed is zero), and index multipliers. 
Uses Mohn's rho for spawning stock biomass as the metric of removing 
retrospective pattern (when ssbrho = 0).
}
