% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_readme.R
\docType{data}
\name{videometa}
\alias{videometa}
\title{Video Meta Data}
\usage{
data(videometa)
}
\description{
Video Meta Data
}
\examples{
##
}
\references{
Institute for Transport Studies, University of Leeds. This work 
was supported by Engineering and Physical Sciences Research Council (EPSRC) 
under Grant number, EP/S005056/1 to Gustav Markkula.
}
\author{
Yi-Shin Lin <yishinlin@pm.me>; Gustav Markkula <g.markkula@leeds.ac.uk>;
Viscando provided the early version of the documentation in trajectory_data_format.pdf (email communications).
}
\keyword{data}
